+++ test/AutoReview/SrcCodeTest.php
@@ -13,6 +13,7 @@ declare(strict_types=1);
 
 namespace Ergebnis\FactoryBot\Test\AutoReview;
 
+use Ergebnis\FactoryGirl\Definition;
 use Ergebnis\Test\Util\Helper;
 use FactoryGirl\Provider\Doctrine;
 use PHPUnit\Framework;
@@ -44,6 +45,9 @@ final class SrcCodeTest extends Framework\TestCase
                 Doctrine\ORM\Locking\TableLockMode::class,
                 Doctrine\ORM\QueryBuilder::class,
                 Doctrine\ORM\Repository::class,
+                Definition\Definitions::class,
+                Definition\Exception\InvalidDefinition::class,
+                Definition\Exception\InvalidDirectory::class,
             ]
         );
     }