+++ CHANGELOG.md
@@ -11,6 +11,7 @@ For a full diff see [`fa9c564...master`][fa9c564...master].
 ### Added
 
 * Imported [`breerly/factory-girl-php@0e6f1b6`](https://github.com/unhashable/factory-girl-php/tree/0e6f1b6724d39108a2e7cef68a74668b7a77b856), ([#1]), by [@localheinz]
+* Imported [`ergebnis/factory-girl-definition@23e57bc`](https://github.com/ergebnis/factory-girl-definition/tree/23e57bc2105ac7a32e3ec7103c866899fe6ad20c), ([#6]), by [@localheinz]
 
 ### Changed
 
@@ -20,5 +21,6 @@ For a full diff see [`fa9c564...master`][fa9c564...master].
 
 [#1]: https://github.com/ergebnis/factory-bot/pull/1
 [#3]: https://github.com/ergebnis/factory-bot/pull/3
+[#6]: https://github.com/ergebnis/factory-bot/pull/6
 
 [@localheinz]: https://github.com/localheinz